agents/python/pyproject.toml
[project]
name = "techtide-agents"
version = "0.1.0"
description = "Python LangGraph/LangChain tools for TechTideAI"
requires-python = ">=3.11"
dependencies = [
  "langchain>=0.2.0",
  "langchain-openai>=0.1.0",
  "langchain-anthropic>=0.1.0",
  "langgraph>=0.2.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.4.0"
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
